Rename worker to athena
