+++ athena/index.js
@@ -1,11 +1,15 @@
 // @flow
+const debug = require('debug')('athena');
 // Our job-processing worker server
 import startMessageNotificationQueue from './queues/message-notification';
 import startReactionNotificationQueue from './queues/reaction-notification';
 import startChannelNotificationQueue from './queues/channel-notification';
 import startCommunityNotificationQueue from './queues/community-notification';
 
+console.log('\n🛠 Athena, the processing worker, is starting...');
+debug('Logging with debug enabled!');
 console.log('');
+
 startMessageNotificationQueue();
 startReactionNotificationQueue();
 startChannelNotificationQueue();
