+++ package.json
@@ -77,18 +77,18 @@
   },
   "scripts": {
     "start": "NODE_ENV=production node build-server/index.js",
-    "start:worker": "NODE_ENV=production node build-worker/index.js",
+    "start:athena": "NODE_ENV=production node build-athena/index.js",
     "start:hermes": "NODE_ENV=production node build-hermes/index.js",
     "dev:client": "NODE_PATH=./ react-app-rewired start",
     "dev:server": "NODE_ENV=development nodemon server/index.js --watch server --exec babel-node",
-    "dev:worker": "NODE_ENV=development nodemon worker/index.js --watch worker --exec babel-node",
+    "dev:athena": "NODE_ENV=development nodemon athena/index.js --watch athena --exec babel-node",
     "dev:hermes": "NODE_ENV=development DEBUG=hermes* nodemon hermes/index.js --watch hermes --exec babel-node",
     "build": "npm run build:server && npm run build:client",
     "prebuild:server": "rimraf build-server",
     "build:server": "NODE_ENV=production babel server --out-dir build-server --ignore 'test/*.js'",
-    "prebuild:worker": "rimraf build-worker",
-    "build:worker": "NODE_ENV=production babel worker --out-dir build-worker --ignore 'test/*.js' --ignore 'node_modules/**/*'",
-    "postbuild:worker": "cp worker/package.json build-worker/package.json && cp ./cacert build-worker/cacert && cp now.json build-worker/now.json",
+    "prebuild:athena": "rimraf build-athena",
+    "build:athena": "NODE_ENV=production babel athena --out-dir build-athena --ignore 'test/*.js' --ignore 'node_modules/**/*'",
+    "postbuild:athena": "cp athena/package.json build-athena/package.json && cp ./cacert build-athena/cacert && cp now.json build-athena/now.json",
     "prebuild:hermes": "rimraf build-hermes",
     "build:hermes": "NODE_ENV=production babel hermes --out-dir build-hermes --ignore 'test/*.js' --ignore 'node_modules/**/*'",
     "postbuild:hermes": "cp hermes/package.json build-hermes/package.json && cp ./cacert build-hermes/cacert && cp now.json build-hermes/now.json",
